Add folder selection functionality to file chooser
The file chooser now supports folder selection in the LWJGL desktop platform. The functionality is introduced through a new enum constant, FOLDER, in the NativeFileChooserIntent class. Adjustments have been made in the DesktopFileChooser class to support this new feature. Finally, a 'Open dir' button has been added to the demo application for showcasing this feature.